+++ src/Terrabuild.Tests/Core/Program.fs
@@ -45,6 +45,7 @@ let private buildGraphNode id projectName projectDir target =
       GraphDef.Node.TargetHash = $"target-{id}"
       GraphDef.Node.ClusterHash = None
       GraphDef.Node.Operations = []
+      GraphDef.Node.EvaluationInputs = []
       GraphDef.Node.Artifacts = GraphDef.ArtifactMode.Workspace
       GraphDef.Node.Build = GraphDef.BuildMode.Auto
       GraphDef.Node.Batch = GraphDef.BatchMode.Single