+++ src/Terrabuild/Core/GraphPipeline/Node.fs
@@ -114,6 +114,7 @@ let build (options: ConfigOptions.Options) (configuration: Configuration.Workspa
                   Node.Target = target
                   Node.Phase = targetConfig.Phase
                   Node.Operations = []
+                  Node.EvaluationInputs = targetConfig.EvaluationInputs
                   Node.Artifacts = targetConfig.Cache |> Option.defaultValue ArtifactMode.Managed
                   Node.Build = build
                   Node.Batch = targetConfig.Batch